third_party/blink/web_tests/virtual/css-text-spacing-trim/external/wpt/css/css-text/parsing/text-spacing-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property text-spacing value 'normal'
  assert_equals: expected "normal" but got "space-first"
[FAIL] Property text-spacing value 'auto'
  assert_true: 'auto' is a supported value for text-spacing. expected true got false
[FAIL] Property text-spacing value 'no-autospace'
  assert_equals: expected "no-autospace" but got "no-autospace space-first"
[FAIL] Property text-spacing value 'trim-start'
  assert_true: 'trim-start' is a supported value for text-spacing. expected true got false
[FAIL] Property text-spacing value 'no-autospace normal'
  assert_true: 'no-autospace normal' is a supported value for text-spacing. expected true got false
[FAIL] Property text-spacing value 'normal no-autospace'
  assert_true: 'normal no-autospace' is a supported value for text-spacing. expected true got false
Harness: the test ran to completion.
